backend/requirements.txt
accounts==0.2.4
alembic==1.8.0
amqp==2.6.1
anyio==3.6.1
api==0.0.7
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
cachetools==5.2.0
celery==4.4.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect==1.1.0
cssutils==2.4.1
Django==3.2
django-cors-headers==3.2.1
django-debug-toolbar==3.4.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
drf-yasg==1.20.0
ecdsa==0.17.0
email-validator==1.2.1
emails==0.5.15
fastapi==0.70.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflect==5.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
lxml==4.9.0
Mako==1.2.0
MarkupSafe==2.0.1
more-itertools==8.13.0
nose==1.3.7
openapi-codec==1.3.2
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2019.3
PyYAML==6.0
raven==6.10.0
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlparse==0.3.0
starlette==0.16.0
tenacity==6.3.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.11.8
uvloop==0.16.0
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
zipp==3.8.0
